+++ transpiler/index.ts
@@ -1,9 +1,0 @@
-const fs = require("fs")
-
-class EHTMLCompound {
-    constructor(directory: string) {
-        /**
-         * TODO: Implement Transpiler
-         */
-    }
-}