+++ CHANGELOG.md
@@ -11,3 +11,14 @@
 * C inclusion
 * Binary Loader
 * Old Blink (Chromium UI)
+
+
+# Version ``0.1.2 beta``
+
+* Stopped development on Protexa embeded Browser (Chromium), Protexa will now only be able to read `.ehtml`
+* Introducing ProtexaEngine, the powerhouse of `.ehtml`
+
+## Deprecations
+
+* `LoadEWebsiteDialog.py`
+* C compiler